+++ Source/ACE.Server/Network/GameAction/Actions/GameActionPutItemInContainer.cs
@@ -1,4 +1,3 @@
-using ACE.Entity;
 
 namespace ACE.Server.Network.GameAction.Actions
 {
@@ -7,8 +6,8 @@ namespace ACE.Server.Network.GameAction.Actions
         [GameAction(GameActionType.PutItemInContainer)]
         public static void Handle(ClientMessage message, Session session)
         {
-            var itemGuid = new ObjectGuid(message.Payload.ReadUInt32());
-            var containerGuid = new ObjectGuid(message.Payload.ReadUInt32());
+            var itemGuid = message.Payload.ReadUInt32();
+            var containerGuid = message.Payload.ReadUInt32();
             var placement = message.Payload.ReadInt32();
 
             session.Player.HandleActionPutItemInContainer(itemGuid, containerGuid, placement);