+++ Source/ACE.Server/Network/GameAction/Actions/GameActionRemoveShortcut.cs
@@ -1,5 +1,3 @@
-using System;
-using ACE.Server.Network.Structure;
 
 namespace ACE.Server.Network.GameAction.Actions
 {