+++ Source/ACE.Server/Network/Structure/AllegianceData.cs
@@ -1,9 +1,9 @@
 using System;
 using System.IO;
 
-using ACE.Entity;
 using ACE.Entity.Enum;
 using ACE.Server.Entity;
+using ACE.Server.Managers;
 using ACE.Server.Network.Enum;
 
 namespace ACE.Server.Network.Structure
@@ -44,7 +44,7 @@ namespace ACE.Server.Network.Structure
             // uint - allegianceAge
 
             // string - name
-            var characterID = new ObjectGuid(0);
+            uint characterID = 0;
             uint cpCached = 0;
             uint cpTithed = 0;
             var bitfield = AllegianceIndex.HasAllegianceAge | AllegianceIndex.HasPackedLevel | AllegianceIndex.LoggedIn;
@@ -62,9 +62,10 @@ namespace ACE.Server.Network.Structure
             if (data.Node != null)
             {
                 var node = data.Node;
-                var player = node.Player;
+                var playerGuid = node.PlayerGuid;
+                var player = PlayerManager.FindByGuid(playerGuid);
 
-                characterID = player.Guid;
+                characterID = player.Guid.Full;
                 gender = (Gender)player.Gender;
                 hg = (HeritageGroup)player.Heritage;
                 rank = (ushort)node.Rank;
@@ -74,7 +75,7 @@ namespace ACE.Server.Network.Structure
                 name = player.Name;
             }
 
-            writer.Write(characterID.Full);
+            writer.Write(characterID);
             writer.Write(cpCached);
             writer.Write(cpTithed);
             writer.Write((uint)bitfield);