+++ Source/ACE.Server/WorldObjects/Player_Use.cs
@@ -36,18 +36,18 @@ namespace ACE.Server.WorldObjects
         // ===============================
         // These are raised by client actions
 
-        public void HandleActionUseWithTarget(ObjectGuid sourceObjectId, ObjectGuid targetObjectId)
+        public void HandleActionUseWithTarget(uint sourceObjectGuid, uint targetObjectGuid)
         {
             StopExistingMoveToChains();
 
-            var invSource = GetInventoryItem(sourceObjectId);
-            var invTarget = GetInventoryItem(targetObjectId);
-            var invWielded = GetEquippedItem(targetObjectId);
+            var invSource = GetInventoryItem(sourceObjectGuid);
+            var invTarget = GetInventoryItem(targetObjectGuid);
+            var invWielded = GetEquippedItem(targetObjectGuid);
 
             if (invSource == null)
             {
                 // is this caster with a built-in spell?
-                var caster = GetEquippedItem(sourceObjectId);
+                var caster = GetEquippedItem(sourceObjectGuid);
                 if (caster != null && caster.SpellDID != null)
                 {
                     // check activation requirements
@@ -58,17 +58,17 @@ namespace ACE.Server.WorldObjects
                         SendUseDoneEvent(WeenieError.SkillTooLow);
                         return;
                     }
-                    HandleActionCastTargetedSpell(targetObjectId, caster.SpellDID ?? 0);
+                    HandleActionCastTargetedSpell(targetObjectGuid, caster.SpellDID ?? 0);
                     return;
                 }
                 else
                 {
-                    log.Warn($"{Name}.HandleActionUseWithTarget({sourceObjectId}, {targetObjectId}): couldn't find {sourceObjectId}");
+                    log.Warn($"{Name}.HandleActionUseWithTarget({sourceObjectGuid:X8}, {targetObjectGuid:X8}): couldn't find {sourceObjectGuid:X8}");
                     SendUseDoneEvent(WeenieError.None);
                 }
             }
 
-            var worldTarget = (invTarget == null) ? CurrentLandblock?.GetObject(targetObjectId) : null;
+            var worldTarget = (invTarget == null) ? CurrentLandblock?.GetObject(targetObjectGuid) : null;
 
             if (invTarget != null)
             {
@@ -102,7 +102,7 @@ namespace ACE.Server.WorldObjects
                 var lp = invSource as Lockpick;
                 lp.HandleActionUseOnTarget(this, worldTarget);
             }
-            else if (targetObjectId == Guid)
+            else if (targetObjectGuid == Guid.Full)
             {
                 // using something on ourselves
                 if (invSource.WeenieType == WeenieType.ManaStone)
@@ -129,19 +129,19 @@ namespace ACE.Server.WorldObjects
             }
         }
 
-        public void HandleActionUseItem(ObjectGuid usedItemId)
+        public void HandleActionUseItem(uint itemGuid)
         {
             StopExistingMoveToChains();
 
             // Search our inventory first
-            var item = GetInventoryItem(usedItemId);
+            var item = GetInventoryItem(itemGuid);
 
             if (item != null)
                 item.UseItem(this);
             else
             {
                 // Search the world second
-                item = CurrentLandblock?.GetObject(usedItemId);
+                item = CurrentLandblock?.GetObject(itemGuid);
 
                 if (item == null)
                 {
@@ -152,7 +152,7 @@ namespace ACE.Server.WorldObjects
                 var moveTo = true;
                 if (item is Container container)
                 {
-                    lastUsedContainerId = usedItemId;
+                    lastUsedContainerId = new ObjectGuid(itemGuid);
 
                     // if the container is already open by this player,
                     // this packet indicates to close the container.